[ADF-4590] - refractor task-cloud service
