+++ lib/process-services-cloud/src/lib/task/start-task/public-api.ts
@@ -16,7 +16,6 @@
  */
 
 export * from './models/task-details-cloud.model';
-export * from './services/start-task-cloud.service';
 export * from './components/start-task-cloud.component';
 export * from './components/people-cloud/people-cloud.component';
 export * from './start-task-cloud.module';
